Reduce compilation time of epee/portable_storage_template_helper.h
