+++ src/checkpoints/checkpoints.cpp
@@ -34,6 +34,8 @@
 #include "string_tools.h"
 #include "storages/portable_storage_template_helper.h" // epee json include
 #include "serialization/keyvalue_serialization.h"
+#include <boost/system/error_code.hpp>
+#include <boost/filesystem.hpp>
 #include <functional>
 #include <vector>
