+++ src/common/i18n.cpp
@@ -35,6 +35,10 @@
 #include "common/i18n.h"
 #include "translation_files.h"
 
+#include <boost/system/error_code.hpp>
+#include <boost/filesystem.hpp>
+#include <algorithm>
+
 #undef MONERO_DEFAULT_LOG_CATEGORY
 #define MONERO_DEFAULT_LOG_CATEGORY "i18n"
