+++ src/wallet/message_store.cpp
@@ -30,6 +30,8 @@
 #include <boost/archive/portable_binary_iarchive.hpp>
 #include <boost/format.hpp>
 #include <boost/algorithm/string.hpp>
+#include <boost/system/error_code.hpp>
+#include <boost/filesystem.hpp>
 #include <fstream>
 #include <sstream>
 #include "file_io_utils.h"
